% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aimPlot.R
\name{aimPlot}
\alias{aimPlot}
\title{Create plot for aims}
\source{
RegData Example data set to run the function
}
\usage{
aimPlot(data, title, size, col1, col2, col3)
}
\arguments{
\item{data}{Data set}

\item{title}{Title for the plot}

\item{size}{Size of the point}

\item{col1}{Colour of the first pie proportion}

\item{col2}{Colour of the second pie proportion}

\item{col3}{Colour of the third pie proportion}
}
\description{
Create a pie like plot to visualize how close a project is to completeness or
achievement of the aims. The middle point show the total completeness of the
aim. The percentages of distribution is 60, 20 and 20. Items for aim should be on
the first row and the percentage of each items is on the second row.
}
\note{
The \code{ggplot2} package is required to run this function
}
\examples{

# basic usage
library("aimPlot")
aimPlot(data = RegData, title = "Plot title")
aimPlot(RegData, "Plot title", 20, "blue", "green", "yellow")


}

