% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RegData}
\alias{RegData}
\title{Sample data}
\format{A data frame with 2 col:
\itemize{
  \item Variable: The specific aims or objectives
  \item Percentage: The percentage of completeness
}}
\usage{
RegData
}
\description{
A sample data for completeness of different aims in a projects. The variables are as follows:
}
\keyword{datasets}

