% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aimPlot.R
\name{aimPlot}
\alias{aimPlot}
\title{Create plot for aims}
\source{
RegData Example datasett to run the function
}
\usage{
aimPlot(data, title)
}
\arguments{
\item{data}{Data set}

\item{title}{Title for the plot}
}
\description{
Create a pie like plot to visualize how close a project is to completeness or
achievement of the aims. The middle point show the total completeness of the
aim. The percentages of distribution is 60, 20 and 20. Items for aim should be on
the first row and the percentage of each items is on the second row.
}
\note{
The \code{ggplot2} package is required to run this function
}
\examples{

# basic usage
library("aimPlot")
aimPlot(data = RegData, title = "Plot title")
aimPlot(RegData, "Plot title")


}

